requirements-api.txt
# Financial Statement Transcription API Requirements

# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI and ML
openai==1.3.7

# PDF Processing
PyMuPDF==1.23.8
Pillow==10.1.0
pdf2image==1.16.3

# Data Processing
pydantic==2.5.0

# Utilities
python-dotenv==1.0.0

# Optional: Enhanced PDF processing
# pdf2image==1.16.3  # Uncomment if you want to use pdf2image instead of PyMuPDF

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing FastAPI endpoints
